add new nodejs script to create sample signatures
