+++ lib/src/test/resources/fixtures/generate_private.js
@@ -3,14 +3,13 @@
  * https://docs.pro.coinbase.com/#private
  */
 
-let crypto = require('crypto');
 let request = require('request');
 let fs = require('fs');
 let uuid = require('uuid');
+let shared = require('./shared');
 
 // https://public.sandbox.pro.coinbase.com/
 let ENDPOINT = "https://api-public.sandbox.pro.coinbase.com";
-let secret = process.env.COINBASE_API_SECRET;
 
 let defaultProductId = "BTC-USD";
 
@@ -20,19 +19,12 @@ let common_headers = {
     'CB-ACCESS-PASSPHRASE': process.env.COINBASE_API_PASSPHRASE
 }
 
-function sign(timestamp, method, requestPath, body) {
-    let what = timestamp + method + requestPath + body;
-    let key = Buffer(secret, 'base64');
-    let hmac = crypto.createHmac('sha256', key);
-    return hmac.update(what).digest('base64');
-}
-
 function get_headers(method, requestPath, body) {
     let timestamp = Date.now() / 1000;
     let encodedBody = body === '' ? body : JSON.stringify(body);
     return Object.assign({}, common_headers, {
         'CB-ACCESS-TIMESTAMP': timestamp,
-        'CB-ACCESS-SIGN': sign(timestamp, method, requestPath, encodedBody)
+        'CB-ACCESS-SIGN': shared.sign(timestamp, method, requestPath, encodedBody)
     });
 }
